pyproject.toml
[project]
name = "clip-finder-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    'requests',
    'Pillow',
    'transformers',
    'numpy<2',
    'load-dotenv',
    'mobileclip@git+https://github.com/ialhashim/ml-mobileclip.git',
    'Pillow>=10.0.0',
    'platformdirs>=3.0.0',
    'osxmetadata',
    'fastapi',
    'uvicorn',
    'pydantic',
    'MulticoreTSNE'
]

[tool.setuptools]
py-modules = ['clip_finder_backend']
